requirements.txt
gradio
ultralytics
opencv-python
pillow
numpy
